+++ src/main/java/com/howtoketocook/ketoapi/response/RecipeResponse.java
@@ -14,7 +14,7 @@ public class RecipeResponse {
 	private String name;
 	private long cookingTime;
 	private long prepTime;
-	private int portions;
+	private long portions;
 	private List<Ingredient> ingredients;
 	private String instructions;
 	private String imageUrl;
